+++ ServersideQoL.AutoStore/CHANGELOG.md
@@ -1,3 +1,7 @@
+### v2.0.13
+- Added option `AutoPickupExcludeItemTypes` to exclude certain item types from auto pickup [#232](https://github.com/ArgusMagnus/ValheimServersideQoL/issues/232)
+- Added option `StackInventoryIntoContainersExcludeItemTypes` to exclude certain item types from stacking into nearby containers [#232](https://github.com/ArgusMagnus/ValheimServersideQoL/issues/232)
+
 ### v2.0.11
 - Moved `AutoPickupMaxRange` config option from ContainerSigns to AutoStore
 - Support for the new core options `UnifiedConfig` and `ConfigPerWorld`